remove dynamic scope check in letrec-hook, propagate scope
I noticed that the evaluator was attempting to resolve dynamic let bindings in the global scope instead of the compiler-letrec-scope.
I also noticed that `invokeUnbound` and `resolveUnbound` operate on a scope but do not take it as a parameter. So I added a `scope` argument to both functions and the corresponding places where they are called. Finally I removed the check which conditionally deletes the scope lines in the compiler hook for `letrec`.
Although the unit tests still pass, I'm not sure whether this solution could have any unintended consequences on performance, and suspect that this condition could be detected some other way, and thus avoid emitting these lines unnecessarily.